Fix a bunch of stuff so that classes appear and can be run
