+++ .idea/misc.xml
@@ -7,5 +7,5 @@
       </list>
     </option>
   </component>
-  <component name="ProjectRootManager" version="2" languageLevel="JDK_1_8" default="true" project-jdk-name="1.8" project-jdk-type="JavaSDK" />
+  <component name="ProjectRootManager" version="2" languageLevel="JDK_1_3" default="false" project-jdk-name="IntelliJ IDEA Community Edition IC-171.3780.107" project-jdk-type="IDEA JDK" />
 </project>
