+++ pom.xml
@@ -4,10 +4,9 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>com.test.school</groupId>
-    <artifactId>homeWork</artifactId>
+    <groupId>com.test.automation</groupId>
+    <artifactId>automationSchool</artifactId>
     <version>1.0-SNAPSHOT</version>
-
 
     <dependencies>
         <dependency>
